Add files via upload
MB had some updates, but it is still messy, and it is out of order.
regsPrep aso updated. 1. code updated to better clean the data. a few for loops updated to better pull the code. It also incorporates the order that MB should be in. Eventually, this document will replace MB.